# skip buf initialization