Since later on, only the bytes that were written into `buf` are read out.
https://github.com/dom96/httpbeast/blob/283e405c1e4a47265b1865638ed49b77b55bb7e5/src/httpbeast.nim#L204-L207