@FlamestoN 2021-06-04
@dom96
Hello!
I don’t know how else to contact you, I didn’t find your mail
May I ask you to post the source code of your c ++ obfuscator? (https://picheta.me/obfuscator)
Or is it too secret? ;)