@dom96 2021-06-05
@FlamestoN I've been keeping it closed source on purpose for a while. Why are you looking for the source? :)