@n5m 2021-06-21
## Benchmark
Ran the server as follows:
```console
$ git checkout <branch>
$ $ nim c --run --gc:arc --hints:off --warnings:off tests/benchmark.nim
Starting 4 threads
Listening on port 8080
```

From another terminal, we send requests to `http://localhost:8080/404` so that the response will have the fewest bytes of the three options, since the response is just `Http404`: https://github.com/dom96/httpbeast/blob/d679e50578969f116cd1bd989e39951535a4d5bf/tests/benchmark.nim#L7-L16

### Without `-d:release -d:danger`
Using the `master` branch (d679e50578969f116cd1bd989e39951535a4d5bf):
```console
$ sudo apt install wrk
$ wrk --threads 3 --connections 500 --duration 300s http://localhost:8080/404
Running 5m test @ http://localhost:8080/404
  3 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.86ms    2.20ms 115.18ms   78.19%
    Req/Sec    55.71k     3.55k   70.98k    69.65%
  49882189 requests in 5.00m, 5.34GB read
  Non-2xx or 3xx responses: 49882189
Requests/sec: 166241.69
Transfer/sec:     18.23MB
```

Using the changes from this change (1b6a3c9d9dfcb6fa462e4bbda9264ed69d07fdfe):
```console
$ wrk --threads 3 --connections 500 --duration 300s http://localhost:8080/404
Running 5m test @ http://localhost:8080/404
  3 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.88ms    2.35ms 154.29ms   83.05%
    Req/Sec    55.96k     3.72k   86.98k    70.29%
  50109635 requests in 5.00m, 5.37GB read
  Non-2xx or 3xx responses: 50109635
Requests/sec: 166980.07
Transfer/sec:     18.31MB
```

### With `-d:release -d:danger`
Using the `master` branch (d679e50578969f116cd1bd989e39951535a4d5bf):
```console
$ wrk --threads 3 --connections 500 --duration 300s http://localhost:8080/404
Running 5m test @ http://localhost:8080/404
  3 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.42ms    4.56ms  95.58ms   83.66%
    Req/Sec    74.58k    22.27k  117.83k    67.67%
  66627327 requests in 5.00m, 7.14GB read
  Non-2xx or 3xx responses: 66627327
Requests/sec: 222032.49
Transfer/sec:     24.35MB
```

Using the changes from this change (1b6a3c9d9dfcb6fa462e4bbda9264ed69d07fdfe):
```console
$ wrk --threads 3 --connections 500 --duration 300s http://localhost:8080/404
Running 5m test @ http://localhost:8080/404
  3 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.38ms    4.56ms 100.59ms   83.74%
    Req/Sec    74.44k    22.09k  118.84k    67.62%
  66493434 requests in 5.00m, 7.12GB read
  Non-2xx or 3xx responses: 66493434
Requests/sec: 221604.85
Transfer/sec:     24.30MB
```

## Summary
### Without `-d:release -d:danger`
This change saw a 0.444% increase in requests per second, from 166.2k req/s to 167.0k req/s.

### With `-d:release -d:danger`
This change saw a 0.193% decrease in requests per second, from 222.0k req/s to 221.6k req/s.

@dom96 Do you have any suggestions to improve the benchmark? I cannot explain why this change would actually degrade performance, but if the benchmark is to be trusted, then perhaps this PR should not be accepted.