@dom96 2021-06-21
@n5m so first I would benchmark a non-404 case, just to avoid any weirdness with wrk when it comes to handling non-2xx responses. But yes, unfortunately this is the tricky thing about working on this project, often you can get very different numbers between runs (even when the server doesn't change). The only thing I can suggest is to run multiple benchmarks and take an average to avoid this, but even looking at these numbers it appears this patch doesn't change things significantly enough.