@dom96 2021-10-29
Will close this as it doesn't seem to bring significant benefits.